Revert "fixed all norm errors"
This reverts commit 4e074363df8472244093f1671e503853b4d4932b.
