+++ philo/Makefile
@@ -1,11 +1,11 @@
 NAME = philo
 
-SRC = main.c arg_parsing.c data.c error.c grim_reaper_utils.c time.c \
+SRC = main.c arg_parsing.c data.c deployment_utils.c error.c time.c \
 	deployment_of_philos.c grim_reaper.c philosopher.c \
 	philosopher_utils.c philosopher_fork_handling.c memory_cleanup.c
 
 CC = cc
-FLAGS = -Wall -Wextra -Werror -O3
+FLAGS = -Wall -Wextra -Werror -g
 DFLAGS = -Wall -Wextra -Werror -g# -fsanitize=address -static-libsan
 PRINT_FLAG = -D PRINT_NUMBER_OF_EATS=1
